Move test to models predictions
