+++ requirements/requirements.test.integration.txt
@@ -5,4 +5,3 @@ pillow
 requests_toolbelt
 numpy<=1.26.4
 pytest-retry<=1.6.3
-supervision~=0.22
