+++ tests/inference/models_predictions_tests/assets/sam2_multipolygon_response.json
